+++ libs/core/src/handlers/commands.ts
@@ -11,6 +11,7 @@ export enum SeelenCommand {
   OpenFile = "open_file",
   SelectFileOnExplorer = "select_file_on_explorer",
   Run = "run",
+  SimulatePerm = "simulate_perm",
   IsDevMode = "is_dev_mode",
   IsAppxPackage = "is_appx_package",
   HasFixedRuntime = "has_fixed_runtime",